requirements.txt
Django==1.11.4
django-cors-headers==2.1.0
django-extensions==1.9.0
django-filter==1.0.4
django-rest-framework==0.1.0
djangorestframework==3.6.4
graphviz==0.8
olefile==0.44
Pillow==4.2.1
pydotplus==2.0.2
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
